+++ zss/gadget/components/tape/elements/activeitem.tsx
