+++ zss/gadget/components/tape/elements/common.ts
