+++ zss/gadget/components/tape/elements/item.tsx
